requirements.txt
horovod==0.28.1
matplotlib==3.8.2
numpy==2.0.1
PyYAML==6.0.1
PyYAML==6.0.2
setuptools==69.0.3
tensorflow>=2.15.0
